Move rootless directory handling to the libpod/pkg/util directory
This should allow us to share this code with buildah.
